+++ generators/ci-cd/command.js
@@ -1,5 +1,5 @@
 import { asCommand } from 'generator-jhipster';
 
 export default asCommand({
-  options: {},
+  configs: {},
 });
