+++ generators/node-server/generators/bootstrap/command.js
