+++ generators/node-server/generators/bootstrap/generator.spec.js
@@ -2,10 +2,10 @@ import { beforeAll, describe, expect, it } from 'vitest';
 
 import { defaultHelpers as helpers, result } from 'generator-jhipster/testing';
 
-const SUB_GENERATOR = 'bootstrap-application';
+const SUB_GENERATOR = 'node-server:bootstrap';
 const SUB_GENERATOR_NAMESPACE = `jhipster-nodejs:${SUB_GENERATOR}`;
 
-describe('SubGenerator bootstrap-application of nodejs JHipster blueprint', () => {
+describe('SubGenerator node-server:bootstrap of nodejs JHipster blueprint', () => {
   describe('run', () => {
     beforeAll(async function () {
       await helpers
